Fixing warnings with clang.

+++ srcs/noyau.c
@@ -8,6 +8,7 @@
 
 noyau_t *creer_noyau(uint32_t dim) {
   noyau_t *core = NULL;
+
   if (dim) {
     core = malloc(sizeof(noyau_t));
     if (core) {
@@ -42,6 +43,7 @@ void read_core(FILE *file, noyau_t *pn) {
   for (size_t y = 0, x = 0; y < pn->dim; y++, x = 0) {
     for (; x < pn->dim; x++) {
       const size_t cur = x + y * pn->dim;
+
       fscanf(file, "%d", &pn->coeffs[cur]);
     }
   }
@@ -55,7 +57,8 @@ noyau_t *charger_noyau(const char *nom_fichier) {
 
     if (file) {
       uint32_t dim = 0;
-      (void)fscanf(file, "%u", &dim);
+
+      (void) fscanf(file, "%u", &dim);
       core = creer_noyau(dim);
       if (core) {
         read_core(file, core);
@@ -65,8 +68,7 @@ noyau_t *charger_noyau(const char *nom_fichier) {
     }
     fclose(file);
   } else {
-    fprintf(stderr,
-        "charger_noyau: filename provided is NULL.\n");
+    fprintf(stderr, "charger_noyau: filename provided is NULL.\n");
   }
   return core;
 }
@@ -74,7 +76,7 @@ noyau_t *charger_noyau(const char *nom_fichier) {
 int core_sum(noyau_t *pn) {
   int res = 0;
 
-  for (int x = 0; x < (pn->dim * pn->dim); x++) {
+  for (size_t x = 0; x < (pn->dim * pn->dim); x++) {
     res += pn->coeffs[x];
   }
   return res;
